refactor: changed for unsigned int

+++ Module-07/ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: revieira <revieira@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/07 17:58:53 by revieira          #+#    #+#             */
-/*   Updated: 2024/01/11 15:13:59 by revieira         ###   ########.fr       */
+/*   Updated: 2024/01/11 16:03:49 by revieira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include <iostream>
 
 template <typename T>
-void	iter(T *array, size_t size, void f(T&))
+void	iter(T *array, unsigned int size, void f(T&))
 {
 	if (array)
 		for (size_t i = 0; i < size; i++)
